Fixes and enhancements to gUnit
Fixes a bug where a testblock that returned instead of
raising an exception was not caught as failing.  Adds
mentioning(_)and(_) variant of assert(_)shouldRaise(_),
so that two separate parts of the error message can be checked.
Eliminates meaningless ", line 0" from description of an object that
cannot be attributed a line number.