Moves dialect-checking into identifierresolution
`identifierresolution.setupContext` already did the external module
check for the dialect, so this check was being done twice.  Now that
`standard` imports many other modules, this was wasteful.
The method `setupContext` now does the dialectCheck, which does the
externalModuleCheck, as well as initializing state to run the
parseChecker and astChecker, if required.