Deploys prefixTree in identifierResolution
Changes name of methods to public, because only
public methods should be described there.  Uses prefixTree
to limit what is written to those names that form a prefix
of a path to a fresh method.  Can't yet limit the menthods
to those that are fresh, because we loose the freshness info
in the symbol table when we inherit from a fresh method.